# specs/1-obe-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OBE System API
  description: |
    Outcome-Based Education Management System API
    
    This API provides endpoints for:
    - Authentication (JWT, Google OAuth, Email OTP)
    - User management with RBAC
    - Course and outcome management
    - Assessment and result tracking
    - Analytics and reporting
  version: 1.0.0
  contact:
    name: OBE System API Support
    email: support@obesystem.com
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.obesystem.com/api
    description: Production server
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [Admin, Teacher, Student]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Create user (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      summary: Update user (Admin or self)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Users]
      summary: Delete user (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}/profile:
    get:
      tags: [Users]
      summary: Get own profile
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  # Course Management Endpoints
...
  /assessments/{id}/results:
    get:
      tags: [Results]
      summary: Get results for assessment (Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResultResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /students/{id}/results:
    get:
      tags: [Results]
      summary: Get student results (Teacher/Student)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: course_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of student results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResultResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /results:
    post:
      tags: [Results]
      summary: Create result entry (Teacher)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResultRequest'
      responses:
        '201':
          description: Result created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /results/bulk:
    post:
      tags: [Results]
      summary: Bulk result entry (Teacher)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                results:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateResultRequest'
      responses:
        '201':
          description: Results created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
        '403':
          $ref: '#/components/responses/Forbidden'
  /results/{id}:
    put:
      tags: [Results]
      summary: Update result (Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResultRequest'
      responses:
        '200':
          description: Result updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /results/{id}/feedback:
    post:
      tags: [Results]
      summary: Add feedback to result (Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment:
                  type: string
      responses:
        '201':
          description: Feedback added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Analytics Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          $ref: '#/components/schemas/RoleResponse'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    CreateUserRequest:
      type: object
      required: [email, first_name, last_name, role_id]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
        role_id:
          type: integer
    UpdateUserRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        role_id:
          type: integer
        is_active:
          type: boolean
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        total_pages:
          type: integer
...
    CLOResponse:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        code:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
    AssessmentResponse:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: string
        weight:
          type: number
        max_score:
          type: number
        due_date:
          type: string
          format: date-time
        clos:
          type: array
          items:
            $ref: '#/components/schemas/CLOResponse'
        created_at:
          type: string
          format: date-time
    CreateResultRequest:
      type: object
      required: [student_id, assessment_id, clo_id, score, max_score]
      properties:
        student_id:
          type: integer
        assessment_id:
          type: integer
        clo_id:
          type: integer
        score:
          type: number
        max_score:
          type: number
    UpdateResultRequest:
      type: object
      properties:
        score:
          type: number
        max_score:
          type: number
    ResultResponse:
      type: object
      properties:
        id:
          type: integer
        student:
          $ref: '#/components/schemas/UserResponse'
        assessment:
          $ref: '#/components/schemas/AssessmentResponse'
        clo:
          $ref: '#/components/schemas/CLOResponse'
        score:
          type: number
        max_score:
          type: number
        percentage:
          type: number
        feedback:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackResponse'
        created_at:
          type: string
          format: date-time
    FeedbackResponse:
      type: object
      properties:
        id:
          type: integer
        result_id:
          type: integer
        comment:
          type: string
        created_by:
          $ref: '#/components/schemas/UserResponse'
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'